back | update back

+++ backend/rentngo/src/main/java/com/example/rentngo/coucheService/ServicesImpl/AdminServiceImpl.java
@@ -5,8 +5,13 @@ import com.example.rentngo.DAO.repository.AdminRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.example.rentngo.coucheService.Services.AdminService;
-import java.util.List;
+import com.example.rentngo.coucheWeb.DTO.AdminRequestDTO;
 
+import lombok.extern.slf4j.Slf4j;
+
+import java.io.IOException;
+import java.util.List;
+@Slf4j
 @Service
 public class AdminServiceImpl implements AdminService {
     @Autowired
@@ -18,22 +23,35 @@ public class AdminServiceImpl implements AdminService {
     }
 
     @Override
-    public Admin addAdmin(Admin admin) {
-        return adminRepository.save(admin);
+    public void addAdmin(AdminRequestDTO adminRequestDTO) throws IOException {
+        Admin admin = new Admin();
+        admin.setUsername(adminRequestDTO.getUsername());
+        admin.setMotdepasse(adminRequestDTO.getMotdepasse());
+        adminRepository.save(admin);
+        log.info("admin added successfully");
     }
 
     @Override
-    public Admin updateAdmin(Long id, Admin adminDetails) {
-        Admin admin = adminRepository.findById(id).orElseThrow();
-        admin.setUsername(adminDetails.getUsername());
-        admin.setMotdepasse(adminDetails.getMotdepasse());
-        return adminRepository.save(admin);
+    public void updateAdmin(AdminRequestDTO adminRequestDTO, Long id) throws IOException {
+        Admin admin = adminRepository.findById(id).orElse(null);
+        if (admin != null) {
+            admin.setUsername(adminRequestDTO.getUsername());
+            admin.setMotdepasse(adminRequestDTO.getMotdepasse());
+            adminRepository.save(admin);
+            log.info("update admin with id: {} done!-------------------------", id);
+        } else {
+            log.warn("Attempted to update a admin with id: {} that does not exist-----------------------", id);
+        }
     }
 
     @Override
     public void deleteAdmin(Long id) {
-        Admin admin = adminRepository.findById(id).orElseThrow();
-        adminRepository.delete(admin);
+        Admin admin = adminRepository.findById(id).orElse(null);
+        if (admin != null) {
+            adminRepository.delete(admin);
+        } else {
+            throw new RuntimeException("Admin with id " + id + " not found");
+        }
     }
 
     @Override
@@ -41,4 +59,3 @@ public class AdminServiceImpl implements AdminService {
         return adminRepository.findByUsername(username);
     }
 }
-
